requirements.txt
pillow
svgwrite
numpy
